Moved CustomName and Material Filter
